Reimplement remaining commands (except `which`)
