+++ commands/exit.js
@@ -1,0 +1,3 @@
+export default function () {
+  window.close()
+}
